refactor(#2389): renamed classes
