+++ eo-maven-plugin/src/main/java/org/eolang/maven/ProbeMojo.java
@@ -42,9 +42,9 @@ import org.eolang.maven.hash.ChRemote;
 import org.eolang.maven.hash.CommitHash;
 import org.eolang.maven.name.ObjectName;
 import org.eolang.maven.name.OnCached;
-import org.eolang.maven.name.OnDefault;
-import org.eolang.maven.name.OnSwap;
 import org.eolang.maven.name.OnVersioned;
+import org.eolang.maven.name.OnSwap;
+import org.eolang.maven.name.OnReplaced;
 import org.eolang.maven.objectionary.Objectionaries;
 import org.eolang.maven.objectionary.ObjsDefault;
 import org.eolang.maven.tojos.ForeignTojo;
@@ -129,7 +129,7 @@ public final class ProbeMojo extends SafeMojo {
                 new ChNarrow(
                     new OnSwap(
                         this.withVersions,
-                        new OnDefault(tojo.identifier(), this.hsh)
+                        new OnVersioned(tojo.identifier(), this.hsh)
                     ).hash()
                 )
             ).withProbed(count);
@@ -166,8 +166,8 @@ public final class ProbeMojo extends SafeMojo {
                 obj -> new OnCached(
                     new OnSwap(
                         this.withVersions,
-                        new OnDefault(
-                            new OnVersioned(ProbeMojo.noPrefix(obj), this.hashes),
+                        new OnVersioned(
+                            new OnReplaced(ProbeMojo.noPrefix(obj), this.hashes),
                             this.hsh
                         )
                     )
