+++ eo-maven-plugin/src/test/java/org/eolang/maven/name/OnUnversionedTest.java
@@ -37,7 +37,7 @@ final class OnUnversionedTest {
     @Test
     void returnsFullNameWithVersions() {
         final String stdout = "stdout";
-        final String object = String.join(OnVersioned.DELIMITER, stdout, "1234567");
+        final String object = String.join(OnReplaced.DELIMITER, stdout, "1234567");
         MatcherAssert.assertThat(
             String.format(
                 "Unversioned object %s as string should have been equal to %s, but it didn't",
@@ -45,7 +45,7 @@ final class OnUnversionedTest {
                 stdout
             ),
             new OnUnversioned(
-                new OnDefault(object, new CommitHash.ChConstant("abcdefg"))
+                new OnVersioned(object, new CommitHash.ChConstant("abcdefg"))
             ).toString(),
             Matchers.equalTo(stdout)
         );
